refactor: refactors enums as top level classes
