+++ src/main/kotlin/tech/cordona/zooonline/bootstrap/animal/AnimalBuilder.kt
@@ -1,10 +1,10 @@
 package tech.cordona.zooonline.bootstrap.animal
 
 import com.github.doyaaaaaken.kotlincsv.dsl.csvReader
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.AMPHIBIAN
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.BIRD
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.MAMMAL
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.REPTILE
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.AMPHIBIAN
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.BIRD
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.MAMMAL
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.REPTILE
 import tech.cordona.zooonline.domain.animal.entity.Animal
 import tech.cordona.zooonline.domain.animal.entity.Animal.Gender.MALE
 import tech.cordona.zooonline.domain.animal.entity.structs.HealthStatistics
