+++ src/main/kotlin/tech/cordona/zooonline/bootstrap/taxonomy/Amphibians.kt
@@ -1,9 +1,9 @@
 package tech.cordona.zooonline.bootstrap.taxonomy
 
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Amphibian.FROG
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Amphibian.SALAMANDER
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.AMPHIBIAN
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Phylum.ANIMAL
+import tech.cordona.zooonline.bootstrap.taxonomy.Amphibian.FROG
+import tech.cordona.zooonline.bootstrap.taxonomy.Amphibian.SALAMANDER
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.AMPHIBIAN
+import tech.cordona.zooonline.bootstrap.taxonomy.Phylum.ANIMAL
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.generateTaxonomyUnits
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getChildrenIds
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getSpeciesFrom
@@ -13,7 +13,7 @@ import tech.cordona.zooonline.domain.taxonomy.entity.TaxonomyUnit
 
 object Amphibians {
 
-	private val amphibianTypes = Taxonomy.Amphibian.values().map { enum -> enum.asString }.toList()
+	private val amphibianTypes = Amphibian.values().map { enum -> enum.asString }.toList()
 
 	private val frogSpecies = listOf("Golden Mantella", "Mountain Chicken", "Tomato Frog")
 	private val salamanderSpecies = listOf("Hellbender", "Tiger Salamander", "Mudpuppy")
