+++ src/main/kotlin/tech/cordona/zooonline/bootstrap/taxonomy/Birds.kt
@@ -1,12 +1,12 @@
 package tech.cordona.zooonline.bootstrap.taxonomy
 
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Bird.CRANE
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Bird.GULL
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Bird.OWL
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Bird.PERCHING
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Bird.PREDATOR
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.BIRD
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Phylum.ANIMAL
+import tech.cordona.zooonline.bootstrap.taxonomy.Bird.CRANE
+import tech.cordona.zooonline.bootstrap.taxonomy.Bird.GULL
+import tech.cordona.zooonline.bootstrap.taxonomy.Bird.OWL
+import tech.cordona.zooonline.bootstrap.taxonomy.Bird.PERCHING
+import tech.cordona.zooonline.bootstrap.taxonomy.Bird.PREDATOR
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.BIRD
+import tech.cordona.zooonline.bootstrap.taxonomy.Phylum.ANIMAL
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.generateTaxonomyUnits
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getChildrenIds
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getSpeciesFrom
@@ -16,7 +16,7 @@ import tech.cordona.zooonline.domain.taxonomy.entity.TaxonomyUnit
 
 object Birds {
 
-	private val birdTypes = Taxonomy.Bird.values().map { enum -> enum.asString }.toList()
+	private val birdTypes = Bird.values().map { enum -> enum.asString }.toList()
 
 	private val craneSpecies = listOf("Sarus Crane", "Stanley Crane", "White-naped Crane")
 	private val gullSpecies = listOf("Cape Thick-knee", "Horned Puffin", "Tufted Puffin")
