+++ src/main/kotlin/tech/cordona/zooonline/bootstrap/taxonomy/Insects.kt
@@ -1,9 +1,9 @@
 package tech.cordona.zooonline.bootstrap.taxonomy
 
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Group.INSECT
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Insect.SCORPION
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Insect.SPIDER
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Phylum.ANIMAL
+import tech.cordona.zooonline.bootstrap.taxonomy.Group.INSECT
+import tech.cordona.zooonline.bootstrap.taxonomy.Insect.SCORPION
+import tech.cordona.zooonline.bootstrap.taxonomy.Insect.SPIDER
+import tech.cordona.zooonline.bootstrap.taxonomy.Phylum.ANIMAL
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.generateTaxonomyUnits
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getChildrenIds
 import tech.cordona.zooonline.bootstrap.taxonomy.TaxonomyUtils.getSpeciesFrom
@@ -13,7 +13,7 @@ import tech.cordona.zooonline.domain.taxonomy.entity.TaxonomyUnit
 
 object Insects {
 
-	private val insectTypes = Taxonomy.Insect.values().map { enum -> enum.asString }.toList()
+	private val insectTypes = Insect.values().map { enum -> enum.asString }.toList()
 
 	private val spiderSpecies = listOf("Black widow", "Brown widow", "Missouri tarantula")
 	private val scorpionSpecies = listOf("Bark scorpion", "Emperor scorpion", "Whip scorpion")
