+++ src/main/kotlin/tech/cordona/zooonline/bootstrap/taxonomy/Taxonomy.kt
@@ -1,8 +1,8 @@
 package tech.cordona.zooonline.bootstrap.taxonomy
 
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Domain.EUKARYOTE
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Kingdom.ANIMALIA
-import tech.cordona.zooonline.bootstrap.taxonomy.Taxonomy.Phylum.ANIMAL
+import tech.cordona.zooonline.bootstrap.taxonomy.Domain.EUKARYOTE
+import tech.cordona.zooonline.bootstrap.taxonomy.Kingdom.ANIMALIA
+import tech.cordona.zooonline.bootstrap.taxonomy.Phylum.ANIMAL
 import tech.cordona.zooonline.domain.taxonomy.entity.TaxonomyUnit
 
 object Taxonomy {
@@ -22,58 +22,58 @@ object Taxonomy {
 
 	val kingdom = TaxonomyUnit(ANIMALIA.asString, EUKARYOTE.asString, mutableSetOf(phylum.id))
 	val domain = TaxonomyUnit(EUKARYOTE.asString, "Life", mutableSetOf(kingdom.id))
+}
 
-	enum class Domain(val asString: String) {
-		EUKARYOTE("Eukaryote")
-	}
+enum class Domain(val asString: String) {
+	EUKARYOTE("Eukaryote")
+}
 
-	enum class Kingdom(val asString: String) {
-		ANIMALIA("Animalia")
-	}
+enum class Kingdom(val asString: String) {
+	ANIMALIA("Animalia")
+}
 
-	enum class Phylum(val asString: String) {
-		ANIMAL("Animal")
-	}
+enum class Phylum(val asString: String) {
+	ANIMAL("Animal")
+}
 
-	enum class Group(val asString: String) {
-		MAMMAL("Mammal"),
-		BIRD("Bird"),
-		REPTILE("Reptile"),
-		INSECT("Insect"),
-		AMPHIBIAN("Amphibian")
-	}
+enum class Group(val asString: String) {
+	MAMMAL("Mammal"),
+	BIRD("Bird"),
+	REPTILE("Reptile"),
+	INSECT("Insect"),
+	AMPHIBIAN("Amphibian")
+}
 
-	enum class Mammal(val asString: String) {
-		CARNIVORE("Carnivore"),
-		ELEPHANT("Elephant"),
-		MONKEY("Monkey"),
-		HOOFED("Hoofed"),
-		POUCHED("Pouched")
-	}
+enum class Mammal(val asString: String) {
+	CARNIVORE("Carnivore"),
+	ELEPHANT("Elephant"),
+	MONKEY("Monkey"),
+	HOOFED("Hoofed"),
+	POUCHED("Pouched")
+}
 
-	enum class Bird(val asString: String) {
-		CRANE("Crane"),
-		GULL("Gull"),
-		OWL("Owl"),
-		PERCHING("Perching"),
-		PREDATOR("Predator")
-	}
+enum class Bird(val asString: String) {
+	CRANE("Crane"),
+	GULL("Gull"),
+	OWL("Owl"),
+	PERCHING("Perching"),
+	PREDATOR("Predator")
+}
 
-	enum class Reptile(val asString: String) {
-		ALLIGATOR("Alligator"),
-		CROCODILE("Crocodile"),
-		LIZARD("Lizard"),
-		SNAKE("Snake"),
-		TURTLE("Turtle")
-	}
+enum class Reptile(val asString: String) {
+	ALLIGATOR("Alligator"),
+	CROCODILE("Crocodile"),
+	LIZARD("Lizard"),
+	SNAKE("Snake"),
+	TURTLE("Turtle")
+}
 
-	enum class Insect(val asString: String) {
-		SPIDER("Spider"),
-		SCORPION("Scorpion"),
-	}
+enum class Insect(val asString: String) {
+	SPIDER("Spider"),
+	SCORPION("Scorpion"),
+}
 
-	enum class Amphibian(val asString: String) {
-		FROG("Frog"),
-		SALAMANDER("Salamander"),
-	}
+enum class Amphibian(val asString: String) {
+	FROG("Frog"),
+	SALAMANDER("Salamander"),
 }
